Create a database in RAM, then dump it to file
The results of `map()` weren't needed; replace with `for_each()`.
Also, there was an error where it wouldn't populate the database when
`map()` was used.

Since RAM is faster than writing to disk, first connect to a database
initialised in RAM, write to it, then save it to a file.

Solution