Use more familiar wildcards on CLI
The input from the CLI was passed straight to the database, which isn't
itself a big problem since there's no important data involved, but using
percent signs or underscores would be pretty tedious, as it isn't
similar to any other CLI tool.

Thus, define a set of rules according to which input wildcards or
characters may be converted to database query equivalents.